requirements.txt
streamlit
python-docx
beautifulsoup4
markdown
openai
python-dotenv
tempfile
python-docx